KAFKA-14365: Refactor Fetcher to allow different implementations
Generalize Fetcher into an interface of the same name and rename the
original Fetcher into FetcherThreadSafe. Generalize appropriate call
sites to use the general interface instead of the concrete
implementation.
